requirements.txt
asgiref==3.3.1
Brotli==1.0.9
dj-database-url==0.5.0
Django==3.1.7
django-adminlte3==0.1.6
django-ckeditor==6.0.0
django-heroku==0.3.1
django-js-asset==1.2.2
djangorestframework==3.12.2
gunicorn==20.0.4
Pillow==8.1.0
psycopg2==2.8.6
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
